requirements.txt
manim
streamlit
numpy
matplotlib
scipy
pandas
plotly
Pillow